+++ Implem.Pleasanter/Libraries/Settings/ContractSettings.cs
@@ -10,7 +10,7 @@ namespace Implem.Pleasanter.Libraries.Settings
         public int? Users;
         public long? Sites;
         public long? Items;
-        public long? Attachments;
+        public int? Attachments;
         public bool? Import;
         public bool? Export;
         public bool? Notice;
